Update First Fixed Bugs of id

+++ src/main/java/com/sprk/imagegallery/controller/ImageController.java
@@ -40,14 +40,16 @@ public class ImageController {
     private static String uploadDir = "src/main/resources/static/images";
 
     @GetMapping("/user/image")
-    public String showImageForm(Model model) {
+    public String showImageForm(Model model, HttpSession session) {
         model.addAttribute("imageObj", new ImageModel());
+        session.setAttribute("imageId", 0);
+        session.setAttribute("imageTitle", "");
 
         return "image-form";
     }
 
-    @PostMapping("/user/image/{userId}")
-    public String processImageForm(@PathVariable int userId, @Valid @ModelAttribute("imageObj") ImageModel imageModel,
+    @PostMapping("/user/image")
+    public String processImageForm(@Valid @ModelAttribute("imageObj") ImageModel imageModel,
             BindingResult bindingResult, @RequestParam("imageData") MultipartFile file, HttpSession session)
             throws Exception {
         UserModel userModel = null;
@@ -56,17 +58,20 @@ public class ImageController {
             return "image-form";
         } else {
             String imageTitle = "";
-            int imageId = imageModel.getImageId();
+            int imageId = (int) session.getAttribute("imageId");
+            // System.out.println(imageId);
+            String existingTitle = (String) session.getAttribute("imageTitle");
             if (!file.isEmpty()) {
 
                 if (imageId != 0) {
 
-                    String existingTitle = imageModel.getImageTitle();
                     // This case is for update
                     // if update user pass image then delete old image
                     Path fileNameAndPath = Paths.get(uploadDir, existingTitle);
                     System.out.println(fileNameAndPath);
                     Files.deleteIfExists(fileNameAndPath);
+                    // System.out.println(imageId);
+                    imageModel.setImageId(imageId);
 
                 }
                 // save image
@@ -74,64 +79,79 @@ public class ImageController {
                 Path fileNameAndPath = Paths.get(uploadDir, imageTitle);
                 Files.write(fileNameAndPath, file.getBytes());
             } else {
-                System.out.println("Inside else of file" + imageModel.getImageTitle());
-                imageTitle = imageModel.getImageTitle();
+                if (imageId == 0) {
+                    // When user is inserting record and didn't provide the image
+                    session.setAttribute("msg", "Please Add Image ");
+                    return "image-form";
+
+                }
+
+                // When user is updating record and didn't provide the image
+                imageTitle = existingTitle;
+                imageModel.setImageId(imageId);
 
             }
 
             // Save user info in db with image
-            userModel = userService.findUserById(userId);
-            if (userModel != null) {
-                imageModel.setImageTitle(imageTitle);
-                imageModel.setUserModel(userModel);
+            userModel = (UserModel) session.getAttribute("user");
 
-                ImageModel savedImage = imageService.saveImage(imageModel);
-                if (savedImage != null) {
-                    if (imageId != 0) {
+            imageModel.setImageTitle(imageTitle);
+            imageModel.setUserModel(userModel);
 
-                        session.setAttribute("msg", "Image updated successfully");
-                    } else {
-                        session.setAttribute("msg", "Image saved successfully");
-                    }
+            ImageModel savedImage = imageService.saveImage(imageModel);
+            if (savedImage != null) {
+                if (imageId != 0) {
+
+                    session.setAttribute("msg", "Image updated successfully");
                 } else {
-                    session.setAttribute("msg", "Something bad happen on server");
-
+                    session.setAttribute("msg", "Image saved successfully");
                 }
-                return "redirect:/imagegallery/user/image/images/" + userModel.getUserId();
             } else {
-                return "redirect:/";
+                session.setAttribute("msg", "Something bad happen on server");
+
             }
+            return "redirect:/imagegallery/user/image/images";
+
         }
     }
 
-    @GetMapping("/user/image/images/{userId}")
-    public String showDashboard(@PathVariable int userId, Model model) {
-        List<ImageModel> allImages = imageService.getAllImagesByUser(userService.findUserById(userId));
+    @GetMapping("/user/image/images")
+    public String showDashboard(Model model, HttpSession session) {
+        UserModel sessionUser = (UserModel) session.getAttribute("user");
+        List<ImageModel> allImages = imageService.getAllImagesByUser(sessionUser);
         model.addAttribute("allImages", allImages);
 
         return "image-dashboard";
     }
 
-    @GetMapping("/user/image/delete/{id}/{userId}")
+    @GetMapping("/user/image/delete/{id}")
     // if passed string in path variable program fails
-    public String deleteImage(@PathVariable int id, @PathVariable int userId, HttpSession session) throws Exception {
+    public String deleteImage(@PathVariable int id, HttpSession session) throws Exception {
         // System.out.println(id);
         ImageModel dbImage = imageService.getImageById(id);
         if (dbImage != null) {
-            boolean result = false;
-            result = imageService.deleteImage(dbImage);
-            String existingTitle = dbImage.getImageTitle();
-            if (result) {
-                Path fileNameAndPath = Paths.get(uploadDir, existingTitle);
-                System.out.println(fileNameAndPath);
-                Files.deleteIfExists(fileNameAndPath);
-                session.setAttribute("msg", "Deleted successfully");
+
+            UserModel sessionUserModel = (UserModel) session.getAttribute("user");
+            UserModel imageUserModel = dbImage.getUserModel();
+            if (sessionUserModel.getUserId() == imageUserModel.getUserId()) {
+
+                boolean result = false;
+                result = imageService.deleteImage(dbImage);
+                String existingTitle = dbImage.getImageTitle();
+                if (result) {
+                    Path fileNameAndPath = Paths.get(uploadDir, existingTitle);
+                    System.out.println(fileNameAndPath);
+                    Files.deleteIfExists(fileNameAndPath);
+                    session.setAttribute("msg", "Deleted successfully");
+                }
+            } else {
+                session.setAttribute("msg", "You are not authorized to change someone else data!!");
             }
 
         } else {
             session.setAttribute("msg", "Image with id = " + id + " doesnot exists in our records!!");
         }
-        return "redirect:/imagegallery/user/image/images/" + userId;
+        return "redirect:/imagegallery/user/image/images";
     }
 
     @GetMapping("/user/image/update/{id}")
@@ -139,9 +159,23 @@ public class ImageController {
     public String updateImage(@PathVariable int id, Model model, HttpSession session) throws Exception {
         // System.out.println(id);
         ImageModel dbImage = imageService.getImageById(id);
+
         if (dbImage != null) {
-            model.addAttribute("imageObj", dbImage);
-            return "image-form";
+
+            UserModel sessionUser = (UserModel) session.getAttribute("user");
+            // Fetch the user information from image and check user is valid or passing some
+            // others id
+            UserModel imageUserModel = dbImage.getUserModel();
+            if (sessionUser.getUserId() == imageUserModel.getUserId()) {
+                model.addAttribute("imageObj", dbImage);
+                session.setAttribute("imageId", dbImage.getImageId());
+                session.setAttribute("imageTitle", dbImage.getImageTitle());
+                return "image-form";
+            } else {
+
+                session.setAttribute("msg", "You are not authorized to change someone else data!!");
+                return "redirect:/imagegallery/user/image/images";
+            }
 
         } else {
